+++ src/base/BlockVariationRegister.php
@@ -67,8 +67,27 @@ class BlockVariationRegister
             'cfgs' => [],
             'vars' => [],
             'extras' => [],
+            'is_default' => false,
         ];
         $this->_tempIdentifier = $identifier;
+        return $this;
+    }
+
+    /**
+     * Add option to register a variations as default. 
+     * 
+     * This means that when a cfg or value has no user input data, this value will be used.
+     * 
+     * This implementation does not check whether another variation has the default check,
+     * so make sure that only a single entry can be the default entry.
+     *
+     * @return \luya\cms\base\BlockVariationRegister
+     * @since 1.0.8
+     */
+    public function asDefault()
+    {
+        $this->_variations[$this->_tempIdentifier]['is_default'] = true;
+
         return $this;
     }